+++ R/oncoplot.R
@@ -25,6 +25,7 @@
 #' @param annotationDat If MAF file was read without clinical data, provide a custom \code{data.frame} with a column \code{Tumor_Sample_Barcode} containing sample names along with rest of columns with annotations.
 #' You can specify which columns to be drawn using `clinicalFeatures` argument.
 #' @param pathways Default `NULL`. Can be `auto`, or a two column data.frame/tsv-file with genes and correspoding pathway mappings.`
+#' @param path_order Default `NULL` Manually specify the order of pathways
 #' @param selectedPathways Manually provide the subset of pathway names to be seletced from `pathways`. Default NULL. In case `pathways` is `auto` draws top 3 altered pathways.
 #' @param draw_titv logical Includes TiTv plot. \code{FALSE}
 #' @param showTumorSampleBarcodes logical to include sample names.
@@ -89,7 +90,7 @@ oncoplot = oncoplot = function(maf, top = 20, minMut = NULL, genes = NULL, alter
                                rightBarData = NULL, rightBarLims = NULL,
                                topBarData = NULL, logColBar = FALSE, includeColBarCN = TRUE,
                                clinicalFeatures = NULL, annotationColor = NULL, annotationDat = NULL,
-                               pathways = NULL, selectedPathways = NULL, draw_titv = FALSE,
+                               pathways = NULL, path_order = NULL, selectedPathways = NULL, draw_titv = FALSE,
                                showTumorSampleBarcodes = FALSE, barcode_mar = 4, barcodeSrt = 90, gene_mar = 5,
                                anno_height = 1, legend_height = 4,
                                sortByAnnotation = FALSE, groupAnnotationBySize = TRUE, annotationOrder = NULL,
@@ -120,9 +121,13 @@ oncoplot = oncoplot = function(maf, top = 20, minMut = NULL, genes = NULL, alter
   } else if(!is.null(pathways)){ #If user provides pathway list
 
     if(is(pathways, 'data.frame')){
+      pathways = data.table::copy(pathways)
       colnames(pathways)[1:2] = c('Gene', 'Pathway')
       data.table::setDT(x = pathways)
       pathways = pathways[!duplicated(Gene)][,.(Gene, Pathway)]
+      if(!is.null(selectedPathways)){
+        pathways = pathways[Pathway %in% selectedPathways]
+      }
     }else if(file.exists(pathways)){
       pathways = data.table::fread(file = pathways)
       colnames(pathways)[1:2] = c('Gene', 'Pathway')
@@ -328,6 +333,14 @@ oncoplot = oncoplot = function(maf, top = 20, minMut = NULL, genes = NULL, alter
     if("Unknown" %in% names(temp_dat)){
       temp_dat_ord = c(grep(pattern = "Unknown", x = names(temp_dat), invert = TRUE, value = TRUE), "Unknown")
       temp_dat = temp_dat[temp_dat_ord]
+    }
+
+    if(!is.null(path_order)){
+      path_order = intersect(x = path_order, names(temp_dat) )
+      temp_dat = temp_dat[path_order]
+      if(length(temp_dat) == 0){
+        stop("None of the pathways match the provided order!")
+      }
     }
 
     nm = lapply(seq_along(temp_dat), function(i){